requirements.txt
asgiref==3.5.2
cloudinary==1.29.0
dj3-cloudinary-storage==0.0.6
Django==3.2.15
djangoshortcuts==0.0.5
gunicorn==20.1.0
pytz==2022.2.1
sqlparse==0.4.2
django-recaptcha==3.0.0
